requirements.txt
streamlit>=1.35.0
snowflake-connector-python>=3.10.0
requests>=2.31.0
pandas>=2.0.0
